% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes_options.R
\name{bayes_options}
\alias{bayes_options}
\title{bayes output options.}
\usage{
bayes_options(
  model.form = "dvi",
  dvcs = NULL,
  svcs = NULL,
  coord.names = c("X", "Y"),
  n.batch = 4,
  batch.length = 500,
  accept.rate = 0.43,
  n.omp.threads = 1,
  verbose = TRUE,
  n.report = 100,
  n.burn = 1000,
  n.thin = 40,
  n.chains = 4,
  save.fitted = TRUE,
  ...
)
}
\arguments{
\item{model.form}{Model form for Bayesian small area estimation model. 
Currently, accepted options are "lm" for linear model, "dvi" for domain 
varying intercept model, "dvc" for domain varying coefficient(s) model, "svi"
for space varying intercept model, and "svc" for space varying coefficient(s) 
model. "dvi" and "svi" models have a random effect on the intercept, and 
"dvc" and "svc" have a random effect on the intercept, along with any 
selected predictors from `dvcs` or `svcs`.}

\item{dvcs}{Character vector. Only used when `model.form` is "dvc". Names of
the columns for which coefficients should vary by domain. Implicitly the 
intercept is included and does not need to specified.}

\item{svcs}{Character vector. Only used when `model.form` is "svc". Names of
the columns for which coefficients should vary in space. Implicitly the
intercept is included and does not need to specified.}

\item{coord.names}{Character vector of length 2. Name of columns for the X
and Y coordinates of plots. Default parameters are tuned for coordinates in 
meters. Used for fitting spatial models.}

\item{n.batch}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{batch.length}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{accept.rate}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{n.omp.threads}{Additional argument to be passed to spAbundance::abund 
or spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{verbose}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{n.report}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{n.burn}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{n.thin}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{n.chains}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{save.fitted}{Additional argument to be passed to spAbundance::abund or 
spAbundance::svcAbund. See documentation in `spAbundance` package for 
details.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for 
fitting Bayesian small area estimation models.
}
\description{
Returns a list of user-supplied parameters and parameter values for outputting
multest with custom aesthetics.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

bayes_options()

}
\author{
Grayson W. White
}
\keyword{options}
