% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R, R/checks_desc.R, R/cool_functions.R
\name{check.numeric}
\alias{check.numeric}
\alias{check.logic.vars}
\alias{check.logic}
\alias{check.matchclass}
\alias{check.matchval}
\alias{check.unique}
\alias{checks_desc}
\alias{checkfilenm}
\alias{checknm}
\alias{check.namedlist}
\title{Check Functions}
\usage{
check.numeric(x)

check.logic.vars(varlst, statement, ignore.case = FALSE, returnVars = FALSE)

check.logic(
  x,
  statement,
  filternm = NULL,
  stopifnull = FALSE,
  stopifinvalid = TRUE,
  removeinvalid = FALSE,
  syntax = "R",
  returnvar = FALSE,
  returnpart = FALSE
)

check.matchclass(
  tab1,
  tab2,
  matchcol,
  var2 = NULL,
  tab1txt = NULL,
  tab2txt = NULL
)

check.matchval(
  tab1,
  tab2,
  var1,
  var2 = NULL,
  tab1txt = NULL,
  tab2txt = NULL,
  gui = FALSE,
  stopifmiss = FALSE,
  subsetrows = FALSE
)

check.unique(x, xvar, uniquex, NAname = "Other", addfactors = TRUE)

checkfilenm(fn, outfolder = NULL, ext = NULL, stopifnull = FALSE)

checknm(nm, nmlst, ignore.case = TRUE)

check.namedlist(xlst, checknms = NULL, modetype = "numeric")
}
\value{
The object who was checked, with necessary changes made for the object to
pass the check (i.e. character vector to numeric vector).
}
\description{
Check functions for objects of class numeric, logical statements,
class matching, value matching, file names, names, and named lists
}
\details{
These functions are used to check if a vector is numeric with check.numeric,
to check a logical statement with check.logic, to check if two objects' 
classes match, to check it two object's values match, to check file names,
to check names, and finally, to check named lists.
}
\author{
Tracey S. Frescino
}
\keyword{internal}
