% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_options.R
\name{database_options}
\alias{database_options}
\title{Database options.}
\usage{
database_options(
  dbname = NULL,
  host = NULL,
  port = NULL,
  user = NULL,
  password = NULL,
  schema = NULL,
  ...
)
}
\arguments{
\item{dbname}{String. Name of database.}

\item{host}{String. Name of database host.}

\item{port}{String. Database port.}

\item{user}{String. User name for database access.}

\item{password}{String. Password for database access.}

\item{schema}{String. Name of schema in database.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for saving data.
}
\description{
Returns a list of user-supplied parameters and parameter values for database access.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

savedata_options(outfolder = "path", overwrite_dsn = FALSE)

}
\author{
Tracey S. Frescino
}
\keyword{options}
