% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBtestPostgreSQL.R
\name{DBtestPostgreSQL}
\alias{DBtestPostgreSQL}
\title{Database - Test a SQLite database table.}
\usage{
DBtestPostgreSQL(dbname, dbconnopen = TRUE)
}
\arguments{
\item{dbname}{String. Name of PostgreSQL database.}

\item{dbconnopen}{Logical. If TRUE, the dbconn connection is not closed.}
}
\value{
An S4 object that inherits from DBIConnection via the DBI package. 
For more information, see `help(DBI::dbConnect)`.
}
\description{
Checks a SQLite database.
}
\author{
Tracey S. Frescino
}
\keyword{data}
