% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/re_exports.R
\docType{data}
\name{DEFAULT_NODATA}
\alias{DEFAULT_NODATA}
\title{List of default nodata values by raster data type}
\format{
An object of class \code{list} of length 8.
}
\usage{
DEFAULT_NODATA
}
\description{
List of default nodata values by raster data type
}
\keyword{datasets}
