% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_refs.R
\name{.update_refs}
\alias{.update_refs}
\title{Updates reference tables}
\usage{
.update_refs(write = FALSE)
}
\arguments{
\item{write}{Logical. Should the internal reference tables be overwritten?}
}
\value{
No return value. Called for side effects.
}
\description{
Updates reference tables
}
\author{
Josh Yamamoto
}
