% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_options.R
\name{eval_options}
\alias{eval_options}
\title{List of population tables.}
\usage{
eval_options(
  evalid = NULL,
  evalCur = FALSE,
  evalEndyr = NULL,
  evalAll = FALSE,
  evalType = "VOL",
  measCur = FALSE,
  measEndyr = NULL,
  invyrs = NULL,
  measyrs = NULL,
  allyrs = FALSE,
  ...
)
}
\arguments{
\item{evalid}{Integer. To extract data for a specific evaluation period. See
notes for more information about FIA Evaluations.}

\item{evalCur}{Logical. If TRUE, extract plots with most current FIA
Evalidation for state(s).}

\item{evalEndyr}{Integer. Defining end year of Evaluation (yyyy).}

\item{evalAll}{Logical. All Evaluations in database.}

\item{evalType}{String vector. The type(s) of evaluation of interest ('ALL',
'AREAVOL', 'GRM', 'P2VEG', 'DWM', 'INV', 'REGEN', 'CRWN').  The evalType
'ALL' includes nonsampled plots; 'AREAVOL' includes plots used for area or
tree estimates (eval_typ %in% c(CURR, VOL)); The evalType 'GRM' includes
plots used for growth, removals, mortality, and change estimates (eval_typ
%in% c(GROW, MORT, REMV, CHNG)). Multiple types are accepted.  See FIA
database manual for regional availability and/or differences.}

\item{measCur}{Logical. If TRUE, extract plots with most current measurement
for state(s).}

\item{measEndyr}{Integer year (YYYY). If measCur=TRUE, extract plots with
most current measurement for state(s) for years measured before measEndyr.}

\item{invyrs}{Integer vector. Defining specific inventory years of data
(e.g., 2010:2015).}

\item{measyrs}{Integer vector. Defining specific measurement years of data
(e.g., 2010:2015).}

\item{allyrs}{Logical. If TRUE, selects all years (annual inventory) in
database.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for data 
evaluation (FIA or custom) extraction to be supplied to *DB functions.
}
\details{
If no parameters, an empty list is returned.
}
\examples{
eval_options(measCur = TRUE)
}
\author{
Tracey S. Frescino
}
\keyword{list}
