% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{GDT_NAMES}
\alias{GDT_NAMES}
\title{Reference tables - gdal data types.}
\format{
A vector of 12 data type values.
}
\source{
gdal values.
}
\description{
Table with gdal data type names.
}
\keyword{datasets}
