% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{ref_conversion}
\alias{ref_conversion}
\title{Reference table - for conversion factors.}
\format{
A dataframe with 6 columns: TYPE, ENGLISH, ENGLISH_ABBR, METRIC, METRIC_ABBR, CONVERSION.
}
\source{
Conversion table.
}
\description{
Table with conversion factors from English to metric units.
}
\keyword{datasets}
