% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{ref_estimators}
\alias{ref_estimators}
\title{Reference table - FIESTA estimators.}
\format{
A dataframe with 7 columns: ESTIMATOR, SHORTNAME, PACKAGE, OUTNAME, ESTIMATOR_TYPE, ESTIMATOR_DATA, DESCRIPTION, CITATION
}
\source{
Comma-delimited file.
}
\description{
Table with list of estimators currently in FIESTA.
}
\keyword{datasets}
