% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{ref_estvar}
\alias{ref_estvar}
\title{Reference table - for generating estimates}
\format{
A data frame to use a reference for estimation 
variables and filters.
}
\description{
Data frame with variable names and descriptions
}
\keyword{datasets}
