% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{ref_evaltyp}
\alias{ref_evaltyp}
\title{Reference table - for generating tables.}
\format{
A dataframe with 3 columns, EVAL_TYP_CD, EVAL_TYP, DESCRIPTION.
}
\source{
FIA look-up table.
}
\description{
Table with row/column domain (VARNM) and their pretty names for table output
(TABLENM).
}
\keyword{datasets}
