% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{ref_units}
\alias{ref_units}
\title{Reference table - for variable units.}
\format{
A dataframe with 3 columns: VARIABLE, UNITS, METRICUNITS.
}
\source{
Units table.
}
\description{
Table with units for TREE variables.
}
\keyword{datasets}
