% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spMakeSpatial_options.R
\name{spMakeSpatial_options}
\alias{spMakeSpatial_options}
\title{Make SpatialPoints options}
\usage{
spMakeSpatial_options(xvar = NULL, yvar = NULL, xy.crs = 4269, ...)
}
\arguments{
\item{xvar}{String. Name of variable in xyplt defining x coordinate.}

\item{yvar}{String. Name of variable in xyplt defining y coordinate.}

\item{xy.crs}{PROJ.4 String or CRS object or Integer EPSG code defining
Coordinate Reference System. (e.g., EPSG:4269-Geodetic coordinate system for
North America, NAD83).}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for making
SpatialPoints.
}
\details{
If no parameters, an empty list is returned.
}
\examples{

spMakeSpatial_options()

}
\author{
Grayson W. White
}
\keyword{options}
