% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableIDs.R
\name{tableIDs}
\alias{tableIDs}
\title{List of FIADB table unique IDs.}
\usage{
tableIDs(
  cond = "PLT_CN",
  plt = "CN",
  tree = "PLT_CN",
  seedling = "PLT_CN",
  subplot = "PLT_CN",
  subp_cond = "PLT_CN",
  condid = "CONDID",
  subpid = "SUBP",
  p2veg_subp_structure = "PLT_CN",
  p2veg_subplot_spp = "PLT_CN",
  invasive_subplot_spp = "PLT_CN",
  cond_dwm_calc = "PLT_CN",
  ...
)
}
\arguments{
\item{cond}{String. Unique identifier of plot in cond.}

\item{plt}{String. Unique identifier of plot in plt.}

\item{tree}{String. Unique identifier of plot in tree.}

\item{seedling}{String. Unique identifier of plot in seedling.}

\item{subplot}{String. Unique identifier of plot in subplot.}

\item{subp_cond}{String. Unique identifier of plot in subp_cond.}

\item{condid}{String. Unique identifier of a condition in cond.}

\item{subpid}{String. Unique identifier of a subplot in subplot and subp_cond.}

\item{p2veg_subp_structure}{String. Unique identifier of a plot in p2veg_subp_structure.}

\item{p2veg_subplot_spp}{String. Unique identifier of a plot in p2veg_subplot_spp.}

\item{invasive_subplot_spp}{String. Unique identifier of a plot in invasive_subplot_spp.}

\item{cond_dwm_calc}{String. Unique identifier of a plot in cond_dwm_calc.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied unique identifier of a plot in tables.
}
\description{
Returns a list of user-supplied parameters and parameter values for data 
table unique IDs to be supplied to *pop functions.
}
\details{
If no parameters, an empty list is returned.
}
\author{
Tracey S. Frescino
}
\keyword{list}
