% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2_desc.R, R/write2csv.R,
%   R/write2postgresql.R, R/write2sqlite.R
\name{write2_desc}
\alias{write2_desc}
\alias{write2csv}
\alias{write2postgresql}
\alias{write2sqlite}
\title{Data Writing Functions}
\usage{
write2csv(
  layer,
  outfile = NULL,
  outfolder = NULL,
  outfilenm = NULL,
  outfn.pre = NULL,
  outfn.date = FALSE,
  overwrite = FALSE,
  tabtitle = NULL,
  appendfile = FALSE,
  closefn = TRUE,
  outtxt = NULL,
  gui = FALSE
)

write2postgresql(
  layer,
  dbconn,
  out_name = NULL,
  lowernames = FALSE,
  schema = NULL,
  append_layer = FALSE,
  overwrite = FALSE,
  index.unique = NULL,
  index = NULL,
  index_type = "btree"
)

write2sqlite(
  layer,
  SQLitefn = NULL,
  out_name = NULL,
  lowernames = FALSE,
  gpkg = FALSE,
  outfolder = NULL,
  overwrite = FALSE,
  append_layer = FALSE,
  createnew = FALSE,
  dbconn = NULL,
  dbconnopen = FALSE,
  index.unique = NULL,
  index = NULL
)
}
\value{
No return value, writes data to either a "csv" or "sqlite" file.
}
\description{
These functions allow for data to be written to different formats. Currently,
supported formats are "csv" and "sqlite".
}
\details{
These functions allow for data to be written to different formats.
}
\author{
Tracey S. Frescino
}
\keyword{internal}
