% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xy_options.R
\name{xy_options}
\alias{xy_options}
\title{List of population tables.}
\usage{
xy_options(
  xy.uniqueid = "PLT_CN",
  xvar = NULL,
  yvar = NULL,
  xy.crs = 4269,
  xyjoinid = NULL,
  ...
)
}
\arguments{
\item{xy.uniqueid}{String. Unique identifier of xy.}

\item{xvar}{String. Name of variable in xyplt defining x coordinate.}

\item{yvar}{String. Name of variable in xyplt defining y coordinate.}

\item{xy.crs}{PROJ.4 String or CRS object or Integer EPSG code defining
Coordinate Reference System.}

\item{xyjoinid}{String. Variable in xy to join to plot data. If NULL,
xyjoinid = xy.uniqueid.}

\item{...}{For extendibility.}
}
\value{
A list of user-supplied parameters and parameter values for strata.
}
\description{
Returns a list of user-supplied parameters and parameter values for data 
xyuation (FIA or custom) extraction to be supplied to *DB functions.
}
\details{
If no parameters, an empty list is returned.
}
\examples{
xy_options(xvar="LON", yvar="LAT")
}
\author{
Tracey S. Frescino
}
\keyword{list}
