% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBtestPostgreSQL.R
\name{DBtestPostgreSQL}
\alias{DBtestPostgreSQL}
\title{Database - Test a PostgreSQL database.}
\usage{
DBtestPostgreSQL(
  dbname = NULL,
  host = NULL,
  port = NULL,
  user = NULL,
  password = NULL,
  dbconnopen = FALSE,
  showlist = TRUE,
  ...
)
}
\arguments{
\item{dbname}{String. Name of the database on the host.}

\item{host}{String. Host name.}

\item{port}{String. Port number.}

\item{user}{String. User name.}

\item{password}{String. Password.}

\item{dbconnopen}{Logical. If TRUE, the database connection is returned and not closed.}

\item{showlist}{Logical. If TRUE, prints list of tables in database.}

\item{...}{Additional authentication arguments passed to DBI::dbConnect}
}
\value{
An S4 object that inherits from DBIConnection via the DBI package if
dbconnopen = TRUE, or NULL otherwise. 
For more information, see `help(DBI::dbConnect)`.
}
\description{
Checks a PostgreSQL database.
}
\author{
Tracey S. Frescino
}
\keyword{data}
