% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{DEFAULT_NODATA}
\alias{DEFAULT_NODATA}
\title{Reference tables - NODATA Values.}
\format{
A list of 6 components.
}
\source{
gdal values.
}
\description{
List of NODATA Values based on data type.
}
\keyword{datasets}
