% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIESTAutils-package.R
\docType{data}
\name{ref_units}
\alias{ref_units}
\title{Reference table - for variable units.}
\format{
A dataframe with 4 columns: VARIABLE, UNITS, METRICUNITS, 
WOODLAND, kg2tons.
}
\source{
Units table.
}
\description{
Table with units for TREE variables. The WOODLAND column was added
to identify which variables include woodland species.The kg2tons
column was added to identify which variables are commonly converted
from kilograms to tons in estimation process.
}
\keyword{datasets}
