% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write2_desc.R, R/write2csv.R, R/write2sqlite.R
\name{write2_desc}
\alias{write2_desc}
\alias{write2csv}
\alias{write2sqlite}
\title{Data Writing Functions}
\usage{
write2csv(
  layer,
  outfile = NULL,
  outfolder = NULL,
  outfilenm = NULL,
  outfn.pre = NULL,
  outfn.date = FALSE,
  overwrite = FALSE,
  tabtitle = NULL,
  appendfile = FALSE,
  closefn = TRUE,
  outtxt = NULL,
  gui = FALSE
)

write2sqlite(
  layer,
  SQLitefn,
  out_name = NULL,
  gpkg = FALSE,
  outfolder = NULL,
  overwrite = FALSE,
  append_layer = FALSE,
  createnew = FALSE,
  dbconnopen = FALSE,
  index.unique = NULL,
  index = NULL
)
}
\description{
These functions allow for data to be written to different formats. Currently,
supported formats are "csv", "gdb", and "sqlite".
}
\details{
These functions allow for data to be written to different formats.
}
\author{
Tracey S. Frescino
}
\keyword{internal}
